requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
langgraph==0.2.28
langchain==0.3.1
langchain-groq==0.2.0
langchain-community==0.3.1
chromadb==0.5.15
gitpython==3.1.43
sentence-transformers==3.1.1
radon==6.0.1
python-dotenv==1.0.1
pydantic==2.9.2
aiofiles==24.1.0
httpx==0.27.2
rich==13.8.1
